# openapi_AI_Business_OS.yaml
openapi: 3.1.0
info:
  title: Agentic AI Business Operating System - API Spec
  version: 1.0.0
  description: |
    Kontrak API lengkap untuk AI Company OS (16 agents + CEO Orchestrator).
    Base URL: https://api.aicompanyos.com/api/v1
    Buku menu untuk Frontend & Backend biar nyambung. Import file ini langsung ke Postman / Swagger UI.
  contact:
    name: Product Team
servers:
  - url: https://api.aicompanyos.com/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Local dev
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login - dapatkan JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email, example: owner@brand.com }
                password: { type: string, example: "rahasia123" }
      responses:
        '200':
          description: Login sukses
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string, example: eyJhbGciOiJIUzI1NiJ9... }
                  refresh_token: { type: string }
                  expires_in: { type: integer, example: 3600 }
                  user: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Token tidak valid / expired
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string, example: unauthorized }
              message: { type: string, example: "Token expired" }
  schemas:
    User:
      type: object
      properties:
        id: { type: string, example: usr_123 }
        email: { type: string, example: owner@brand.com }
        name: { type: string, example: Budi }
        role: { type: string, enum: [owner, admin, finance, marketing, strategy, viewer] }
...